+++ src/GridLayout/utils/positionFn.ts
@@ -1,5 +1,6 @@
 import { innerHeight, outerHeight, innerWidth, outerWidth } from './domFns';
 import { int } from './other';
+
 import type {
   Constraint,
   Layout,
