+++ src/Widget/Note/Note.tsx
@@ -1,5 +1,6 @@
 import React, { useRef } from 'react';
 import styled from 'styled-components';
+
 import Title from './Title';
 import Editor from './Editor';
