+++ src/Widget/Note/RestoreContentPlugin.tsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from 'react';
+
 import { EditorState } from 'lexical';
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
 import { OnChangePlugin } from '@lexical/react/LexicalOnChangePlugin';
