+++ src/Widget/Todo/TodoItem.tsx
@@ -1,6 +1,8 @@
 import React, { useRef } from 'react';
 import styled, { css } from 'styled-components';
+
 import ContentEditable, { ContentEditableEvent } from 'react-contenteditable';
+
 import type { TodoData } from './Todo';
 import { Close } from '../../components/Icons';
 import { useHover } from '../../hooks';
