+++ src/Widget/renderWidget.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import type { RenderWidgetData, WidgetData } from '../types/WidgetTypes';
 import Clock from './Clock';
 import Note from './Note';
 import Todo from './Todo';
@@ -7,6 +6,8 @@ import EnglishCard from './EnglishCard';
 import News from './News';
 import Weather from './Weather';
 import Calendar from './Calendar';
+
+import type { RenderWidgetData, WidgetData } from '../types/WidgetTypes';
 
 export default function renderWidget(
   widget: RenderWidgetData,
