+++ src/components/GoogleLoginButton/GoogleLoginButton.tsx
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
+
 import GoogleLogo from './GoogleLogo';
 
 const Wrapper = styled.div`
