+++ src/layout/NavBar/SelectNewWidget/SelectNewWidget.tsx
@@ -1,13 +1,17 @@
 import React, { useRef } from 'react';
+import styled from 'styled-components';
+
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { EffectCoverflow } from 'swiper';
-import styled from 'styled-components';
 import { v4 } from 'uuid';
-import { DoubleArrow } from '../../../components/Icons';
-import { widgetConfig } from '../../../Widget';
-import { SwiperButtonNext, SwiperButtonPrev } from '../../../components/Swiper';
+
 import WidgetSlide from './WidgetSlide';
+import { widgetConfig } from '../../../Widget';
+import { DoubleArrow } from '../../../components/Icons';
+import { SwiperButtonNext, SwiperButtonPrev } from '../../../components/Swiper';
+
 import type { WidgetData, WidgetType } from '../../../types/WidgetTypes';
+
 import 'swiper/css';
 import 'swiper/css/effect-coverflow';
