+++ src/layout/NavBar/Settings.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
+
 import {
   ContentCopy,
   Github,
