+++ src/utils/lib.ts
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/no-cycle
 import type { WidgetData, WidgetType } from '../types/WidgetTypes';
 
 export function getRandomInt(max: number) {
